+++ History.md
@@ -14,6 +14,7 @@
 * Changes behavior of `--keep-line-breaks`/`keepBreaks` option to keep breaks after trailing braces only.
 * Makes all multiple selectors ordered alphabetically (aids merging).
 * Adds property overriding so more coarse properties override more granular ones.
+* Adds reducing non-adjacent selectors.
 
 1.1.7 / 2013-10-28
 ==================